refactor: move server folder, add tsconfig for server and start script
